+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.9.0
 - Features:
-
+  - Add syntax to clear multiple stores at the same time. Use `lunaris.clear('@myCommonStore*')`,
 - Fix:
 
 ## 1.8.0